refactor(ui): move navigation to web package
