+++ packages/ui/index.ts
@@ -16,13 +16,5 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-// Contexts
-export * from './src/authentication/AuthContext';
-
-// UI Components
 export * from './src/style/GlobalStyle';
 export * from './src/button/Button';
-export * from './src/navigation/search/Search';
-export * from './src/navigation/auth-buttons/AuthButtons';
-export * from './src/navigation/Navigation';
-export * from './src/layout/Layout';
