+++ packages/web/src/components/navigation/search/Search.styles.ts
